requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-environ==0.11.2
drf-spectacular==0.27.2
celery==5.4.0
redis==5.2.0
psycopg2-binary==2.9.5
django-environ==0.11.2
requests==2.27.1